requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
